--- test/NuGet.Clients.Tests/NuGet.CommandLine.Test/MockResponses/MockResponse.cs
@@ -1,8 +1,0 @@
-namespace NuGet.CommandLine.Test
-{
-    public class MockResponse
-    {
-        public string ContentType { get; set; }
-        public byte[] Content { get; set; }
-    }
-}
