+++ test/NuGet.Clients.Tests/NuGet.CommandLine.Test/Util.cs
@@ -18,6 +18,7 @@ using NuGet.Packaging;
 using NuGet.Packaging.Core;
 using NuGet.Test.Utility;
 using NuGet.Versioning;
+using Test.Utility;
 using Xunit;
 
 namespace NuGet.CommandLine.Test
@@ -452,81 +453,32 @@ namespace NuGet.CommandLine.Test
 
         public static JObject CreateIndexJson()
         {
-            return JObject.Parse(@"
-{
-    ""version"": ""3.2.0"",
-    ""resources"": [],
-    ""@context"": {
-        ""@vocab"": ""http://schema.nuget.org/services#"",
-        ""comment"": ""http://www.w3.org/2000/01/rdf-schema#comment""
-    }
-}");
+            return FeedUtilities.CreateIndexJson();
         }
 
         public static void AddFlatContainerResource(JObject index, MockServer server)
         {
-            var resource = new JObject
-            {
-                { "@id", $"{server.Uri}flat" },
-                { "@type", "PackageBaseAddress/3.0.0" }
-            };
-
-            var array = index["resources"] as JArray;
-            array.Add(resource);
+            FeedUtilities.AddFlatContainerResource(index, server.Uri);
         }
 
         public static void AddRegistrationResource(JObject index, MockServer server)
         {
-            var resource = new JObject
-            {
-                { "@id", $"{server.Uri}reg" },
-                { "@type", "RegistrationsBaseUrl/3.0.0-beta" }
-            };
-
-            var array = index["resources"] as JArray;
-            array.Add(resource);
+            FeedUtilities.AddRegistrationResource(index, server.Uri);
         }
 
         public static void AddLegacyGalleryResource(JObject index, MockServer serverV2, string relativeUri = null)
         {
-            var resourceUri = new Uri(serverV2.Uri);
-            if (relativeUri != null)
-            {
-                resourceUri = new Uri(resourceUri, relativeUri);
-            }
-
-            var resource = new JObject
-            {
-                { "@id", resourceUri },
-                { "@type", "LegacyGallery/2.0.0" }
-            };
-
-            var array = index["resources"] as JArray;
-            array.Add(resource);
+            FeedUtilities.AddLegacyGalleryResource(index, serverV2.Uri, relativeUri);
         }
 
         public static void AddPublishResource(JObject index, MockServer publishServer)
         {
-            var resource = new JObject
-            {
-                { "@id", $"{publishServer.Uri}push" },
-                { "@type", "PackagePublish/2.0.0" }
-            };
-
-            var array = index["resources"] as JArray;
-            array.Add(resource);
+            FeedUtilities.AddPublishResource(index, publishServer.Uri);
         }
 
         public static void AddPublishSymbolsResource(JObject index, MockServer publishServer)
         {
-            var resource = new JObject
-            {
-                { "@id", $"{publishServer.Uri}push" },
-                { "@type", "SymbolPackagePublish/4.9.0" }
-            };
-
-            var array = index["resources"] as JArray;
-            array.Add(resource);
+            FeedUtilities.AddPublishSymbolsResource(index, publishServer.Uri);
         }
 
         public static void CreateConfigForGlobalPackagesFolder(string workingDirectory)
@@ -627,92 +579,7 @@ namespace NuGet.CommandLine.Test
         /// </summary>
         public static JObject CreateSinglePackageRegistrationBlob(MockServer server, string id, string version)
         {
-            return CreatePackageRegistrationBlob(server, id, new KeyValuePair<string, bool>[] { new KeyValuePair<string, bool>(version, true) });
-        }
-
-        /// <summary>
-        /// Create a registration blob for a package
-        /// </summary>
-        public static JObject CreatePackageRegistrationBlob(MockServer server, string id, IEnumerable<KeyValuePair<string, bool>> versionToListedMap)
-        {
-            var indexUrl = string.Format(CultureInfo.InvariantCulture,
-                                    "{0}reg/{1}/index.json", server.Uri, id);
-            var lowerBound = "0.0.0";
-            var upperBound = "9.0.0";
-            var regBlob = new JObject();
-            regBlob.Add(new JProperty("@id", indexUrl));
-            var typeArray = new JArray();
-            regBlob.Add(new JProperty("@type", typeArray));
-            typeArray.Add("catalog: CatalogRoot");
-            typeArray.Add("PackageRegistration");
-            typeArray.Add("catalog: Permalink");
-
-            regBlob.Add(new JProperty("commitId", Guid.NewGuid()));
-            regBlob.Add(new JProperty("commitTimeStamp", "2015-06-22T22:30:00.1487642Z"));
-            regBlob.Add(new JProperty("count", "1"));
-
-            var pages = new JArray();
-            regBlob.Add(new JProperty("items", pages));
-
-            var page = new JObject();
-            pages.Add(page);
-
-            page.Add(new JProperty("@id", indexUrl + $"#page/{lowerBound}/{upperBound}"));
-            page.Add(new JProperty("@type", indexUrl + "catalog:CatalogPage"));
-            page.Add(new JProperty("commitId", Guid.NewGuid()));
-            page.Add(new JProperty("commitTimeStamp", "2015-06-22T22:30:00.1487642Z"));
-            page.Add(new JProperty("count", versionToListedMap.Count()));
-            page.Add(new JProperty("parent", indexUrl));
-            page.Add(new JProperty("lower", lowerBound));
-            page.Add(new JProperty("upper", upperBound));
-
-            var items = new JArray();
-            page.Add(new JProperty("items", items));
-            foreach (var versionToListed in versionToListedMap)
-            {
-                var item = GetPackageRegistrationItem(server, id, version: versionToListed.Key, listed: versionToListed.Value, indexUrl);
-                items.Add(item);
-            }
-            return regBlob;
-        }
-
-        private static JObject GetPackageRegistrationItem(MockServer server, string id, string version, bool listed, string indexUrl)
-        {
-            var item = new JObject();
-
-            item.Add(new JProperty("@id",
-                    string.Format("{0}reg/{1}/{2}.json", server.Uri, id, version)));
-
-            item.Add(new JProperty("@type", "Package"));
-            item.Add(new JProperty("commitId", Guid.NewGuid()));
-            item.Add(new JProperty("commitTimeStamp", "2015-06-22T22:30:00.1487642Z"));
-
-            var catalogEntry = new JObject();
-            item.Add(new JProperty("catalogEntry", catalogEntry));
-            item.Add(new JProperty("packageContent", $"{server.Uri}flat/{id}/{version}/{id}.{version}.nupkg"));
-            item.Add(new JProperty("registration", indexUrl));
-
-            catalogEntry.Add(new JProperty("@id",
-                string.Format("{0}catalog/{1}/{2}.json", server.Uri, id, version)));
-
-            catalogEntry.Add(new JProperty("@type", "PackageDetails"));
-            catalogEntry.Add(new JProperty("authors", "test"));
-            catalogEntry.Add(new JProperty("description", "test"));
-            catalogEntry.Add(new JProperty("iconUrl", ""));
-            catalogEntry.Add(new JProperty("id", id));
-            catalogEntry.Add(new JProperty("language", "en-us"));
-            catalogEntry.Add(new JProperty("licenseUrl", ""));
-            catalogEntry.Add(new JProperty("listed", listed));
-            catalogEntry.Add(new JProperty("minClientVersion", ""));
-            catalogEntry.Add(new JProperty("projectUrl", ""));
-            catalogEntry.Add(new JProperty("published", "2015-06-22T22:30:00.1487642Z"));
-            catalogEntry.Add(new JProperty("requireLicenseAcceptance", false));
-            catalogEntry.Add(new JProperty("summary", ""));
-            catalogEntry.Add(new JProperty("title", ""));
-            catalogEntry.Add(new JProperty("version", version));
-            catalogEntry.Add(new JProperty("tags", new JArray()));
-
-            return item;
+            return FeedUtilities.CreatePackageRegistrationBlob(server.Uri, id, new KeyValuePair<string, bool>[] { new KeyValuePair<string, bool>(version, true) });
         }
 
         public static string CreateProjFileContent(
