+++ test/NuGet.Clients.Tests/NuGet.CommandLine.Test/XunitAttribute.cs
@@ -40,6 +40,10 @@ namespace NuGet.CommandLine.Test
             {
                 Skip = "Skip this test on mono for now.";
             }
+            else
+            {
+                Skip = null;
+            }
         }
     }
