+++ test/TestUtilities/Test.Utility/MockResponses/MockResponse.cs
@@ -1,0 +1,11 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+namespace Test.Utility
+{
+    public class MockResponse
+    {
+        public string ContentType { get; set; }
+        public byte[] Content { get; set; }
+    }
+}
